goes as far as checking/creating policy and SSO group
