+++ src/cli/lib/check-authentication.mjs
@@ -2,8 +2,8 @@ import { ListBucketsCommand, S3Client } from '@aws-sdk/client-s3'
 
 import { getCredentials } from '../../lib/actions/lib/get-credentials' // move to shared
 
-const checkAuthentication = async ({ db }) => {
-  const credentials = getCredentials(db.account.settings)
+const checkAuthentication = async ({ db } = {}) => {
+  const credentials = getCredentials(db?.account?.settings) // passes in 'sso-profile'
 
   const s3Client = new S3Client({ credentials })
   const listBucketsCommand = new ListBucketsCommand({})
