+++ src/cli/lib/handle-configuration.mjs
@@ -4,6 +4,7 @@ import { cliSpec } from '../constants'
 import { getOptionsSpec } from './get-options-spec'
 import { handleConfigurationLocalSetup } from './configuration/handle-configuration-local-setup'
 import { handleConfigurationShow } from './configuration/handle-configuration-show'
+import { handleConfigurationSSOSetup } from './configuration/handle-configuration-sso-setup'
 
 const handleConfiguration = async ({ argv, db }) => {
   const configurationOptionsSpec = getOptionsSpec({ cliSpec, name : 'configuration' })
@@ -16,6 +17,8 @@ const handleConfiguration = async ({ argv, db }) => {
       return await handleConfigurationLocalSetup({ argv, db })
     case 'show':
       return await handleConfigurationShow({ argv, db })
+    case 'sso-setup':
+      return await handleConfigurationSSOSetup({ argv, db })
     default:
       throw new Error('Unknown configuration command: ' + subcommand)
   }
