+++ src/lib/actions/lib/get-credentials.mjs
@@ -1,6 +1,6 @@
 import { fromIni } from '@aws-sdk/credential-providers'
 
-const getCredentials = ({ 'sso-profile': ssoProfile }) => {
+const getCredentials = ({ 'sso-profile': ssoProfile } = {}) => {
   ssoProfile = ssoProfile || process.env.AWS_PROFILE || 'default'
 
   const credentials = fromIni({ profile : ssoProfile })
